Refactored section A to use OptionInput component
