+++ app/components/SectionAOptions.jsx
@@ -1,79 +1,53 @@
-import { useState } from "react"
 import {
-   FormControl,
-   FormControlLabel,
-   FormLabel
+   Container,
+   Typography,
+   Box
 } from "@mui/material"
-import RadioGroup from "@mui/material/RadioGroup"
-import Radio from "@mui/material/Radio"
-import CheckIcon from "@mui/icons-material/Check"
-import CheckBoxOutlineBlankOutlinedIcon from "@mui/icons-material/CheckBoxOutlineBlankOutlined"
+import OptionInput from "./OptionInput"
+import { sectionAOpts } from "@/lib/options"
 
-export default function CheckOptions({ option }) {
-   const [value, setValue] = useState(false)
-
-   const handleChange = (evt) =>
-      setValue(evt.target.value)
-
+export default function CheckOptions() {
    return (
-      <FormControl
-         className="mb-2 flex flex-nowrap items-center justify-between"
-         sx={{ flexDirection: "row" }}
-      >
-         <FormLabel
-            id={`${option} radio group`}
-            className="text-black"
+      <Container className="mt-6 px-0">
+         <Typography
+            variant="h3"
+            marginBottom={1}
          >
-            {option}
-         </FormLabel>
-         <RadioGroup
-            aria-labelledby={`${option} radio group`}
-            name={`${option} radio group`}
-            value={value}
-            onChange={handleChange}
+            <span className="font-bold">A.</span>{" "}
+            Are Permits Required? Are they
+            displayed and properly signed by the
+            PSC/PSA?
+         </Typography>
+         <Box
             sx={{
-               display: "inline-block"
-            }}
-         >
-            <FormControlLabel
-               value={true}
-               control={
-                  <Radio
-                     checkedIcon={<CheckIcon />}
-                     icon={
-                        <CheckBoxOutlineBlankOutlinedIcon />
-                     }
-                  />
+               display: "flex",
+               flexDirection: {
+                  xs: "column",
+                  tablet: "row"
                }
-               label="Yes"
+            }}
+            paddingX={1}
+         >
+            <Box
                sx={{
-                  visibility:
-                     option === "Other" &&
-                     "hidden",
-                  marginRight: {
-                     xs: "14px",
-                     sm: "10px"
+                  width: {
+                     tablet: "48%"
+                  },
+                  marginLeft: {
+                     tablet: 0
                   }
                }}
-            />
-            <FormControlLabel
-               value={false}
-               control={
-                  <Radio
-                     checkedIcon={<CheckIcon />}
-                     icon={
-                        <CheckBoxOutlineBlankOutlinedIcon />
-                     }
-                  />
-               }
-               label="No"
-               sx={{
-                  visibility:
-                     option === "Other" &&
-                     "hidden"
-               }}
-            />
-         </RadioGroup>
-      </FormControl>
+            >
+               {sectionAOpts.map((opt, idx) => {
+                  return (
+                     <OptionInput
+                        key={idx}
+                        option={opt}
+                     />
+                  )
+               })}
+            </Box>
+         </Box>
+      </Container>
    )
 }
